pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "whisper-voice-to-text"
version = "0.1.0"
authors = [
    {name = "Jenner Torrence", email = "jenner@consiliency.io"},
]
description = "Voice-to-text application using OpenAI's Whisper model"
readme = "README.MD"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    # Core packages
    "numpy>=1.24.3",
    "sounddevice>=0.4.6",
    "torch>=2.1.0",
    "torchaudio>=2.1.0",
    "openai-whisper>=20231117",
    "pyaudio>=0.2.13",
    "pynput>=1.7.6",
    
    # LLM integrations
    "openai>=1.12.0",
    "requests>=2.31.0",
    
    # Optional packages
    "scipy>=1.11.3",
    "python-dotenv>=1.0.0",
    
    # System checks
    "psutil>=5.9.0",
    "packaging>=23.0",
    
    # API clients
    "anthropic>=0.15.0",
    "google-generativeai>=0.3.2",
    
    # Ollama support
    "ollama>=0.1.0",
    "retry>=0.9.2",
    
    # PyTorch dependencies
    "typing-extensions>=4.7.0",
    "filelock>=3.12.0",
    "sympy>=1.12",
    "networkx>=3.1",
    "jinja2>=3.0",
    "pyperclip>=1.8.2",
    # Tray icon support
    # "pystray>=0.19.4",
    # "Pillow>=10.3.0",
    # "win10toast_click>=0.0.3; platform_system=='Windows'",
]

[project.optional-dependencies]
dev = [
    "black>=23.1.0",
    "pytest>=7.3.1",
    "mypy>=1.2.0"
]

[project.urls]
"Homepage" = "https://github.com/yourusername/whisper_voice_to_text"
"Bug Tracker" = "https://github.com/yourusername/whisper_voice_to_text/issues"

[project.scripts]
whisper-voice-to-text = "voicetotext.main:main"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["voicetotext"]

[tool.uv]
index-url = "https://pypi.org/simple"
extra-index-url = ["https://download.pytorch.org/whl/cu121"]
